requirements.txt
# Core Web UI
streamlit>=1.28.0

# Document Parsing
pdfplumber>=0.7.6
python-docx>=0.8.11
PyPDF2>=3.0.0  # Required for fallback PDF parsing

# Google Gemini API
google-generativeai>=0.3.0
python-dotenv>=1.0.0  # For loading API key from .env

# Vector DB + Embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# NLP + ML
nltk>=3.8.1
spacy>=3.6.1
scikit-learn>=1.3.0
textblob>=0.17.1
transformers>=4.30.0
tokenizers>=0.13.0
huggingface-hub>=0.16.0
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
pickle-mixin>=1.0.2
pathlib>=1.0.1
typing-extensions>=4.5.0
chardet>=5.1.0
python-magic>=0.4.27
orjson>=3.9.0
psutil>=5.9.0
regex>=2023.6.3
requests>=2.31.0
urllib3>=2.0.0
argparse>=1.4.0

# Dev and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
rich>=13.4.0
icecream>=2.1.3
aiofiles>=23.1.0

# Optional (for async if needed)
asyncio>=3.4.3
